# Speed up output rendering: add a limit on max length of protocol to linkify